Make utils have one function per file
