+++ tests/e2e/specs/block-editor/typography.test.js
@@ -1,6 +1,6 @@
 import { createNewPost, setPostContent } from '@wordpress/e2e-test-utils';
 import { TPOGRAPHY_TEST_POST_CONTENT } from '../../utils/post';
-import { setCustomize } from '../../utils/customize';
+import { setCustomize } from '../../utils/set-customize';
 
 describe( 'Typography in the block editor', () => {
 	it( 'should work as set in the customizer for h1-h6', async () => {
